requirements.txt
face_recognition
numpy
opencv-python
Pillow
pandas
